+++ cbr/src/main/java/edu/cbr/dto/ConvertResponse.java
@@ -2,6 +2,8 @@ package edu.cbr.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 public record ConvertResponse(
         @Schema(description = "The currency being converted from", example = "USD")
         String fromCurrency,
@@ -10,6 +12,6 @@ public record ConvertResponse(
         String toCurrency,
 
         @Schema(description = "The converted amount", example = "9000.5")
-        Double convertedAmount
+        BigDecimal convertedAmount
 ) {
 }
